refactor: split SimplifiedFlightFormV2.tsx into FlightForm/ step components
